Fix issue with quarkus.init-and-exit
Fixes: #32729
