+++ core/runtime/src/main/java/io/quarkus/runtime/init/InitializationTaskRecorder.java
@@ -1,10 +1,14 @@
 package io.quarkus.runtime.init;
 
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+import java.util.function.Supplier;
 import java.util.stream.StreamSupport;
 
 import org.eclipse.microprofile.config.ConfigProvider;
 
 import io.quarkus.runtime.PreventFurtherStepsException;
+import io.quarkus.runtime.Quarkus;
 import io.quarkus.runtime.annotations.Recorder;
 
 /**
@@ -22,8 +26,27 @@ public class InitializationTaskRecorder {
                 .anyMatch(n -> QUARKUS_INIT_AND_EXIT.equals(n));
         if (initAndExitConfigured) {
             if (ConfigProvider.getConfig().getValue(QUARKUS_INIT_AND_EXIT, boolean.class)) {
-                throw new PreventFurtherStepsException("Gracefully exiting after initalization.", 0);
+                preventFurtherRecorderSteps(5, "Error attempting to gracefully shutdown after initialization",
+                        () -> new PreventFurtherStepsException("Gracefully exiting after initialization.", 0));
             }
         }
+    }
+
+    public static void preventFurtherRecorderSteps(int waitSeconds, String waitErrorMessage,
+            Supplier<PreventFurtherStepsException> supplier) {
+        CountDownLatch latch = new CountDownLatch(1);
+        new Thread(new Runnable() {
+            @Override
+            public void run() {
+                Quarkus.blockingExit();
+                latch.countDown();
+            }
+        }).start();
+        try {
+            latch.await(waitSeconds, TimeUnit.SECONDS);
+        } catch (InterruptedException e) {
+            System.err.println(waitErrorMessage);
+        }
+        throw supplier.get();
     }
 }
