+++ extensions/arc/runtime/src/main/java/io/quarkus/arc/runtime/appcds/AppCDSRecorder.java
@@ -1,31 +1,17 @@
 package io.quarkus.arc.runtime.appcds;
 
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
 import io.quarkus.runtime.PreventFurtherStepsException;
-import io.quarkus.runtime.Quarkus;
 import io.quarkus.runtime.annotations.Recorder;
+import io.quarkus.runtime.init.InitializationTaskRecorder;
 
 @Recorder
 public class AppCDSRecorder {
 
     public void controlGenerationAndExit() {
         if (Boolean.parseBoolean(System.getProperty("quarkus.appcds.generate", "false"))) {
-            CountDownLatch latch = new CountDownLatch(1);
-            new Thread(new Runnable() {
-                @Override
-                public void run() {
-                    Quarkus.blockingExit();
-                    latch.countDown();
-                }
-            }).start();
-            try {
-                latch.await(5, TimeUnit.SECONDS);
-            } catch (InterruptedException e) {
-                System.err.println("Unable to properly shutdown Quarkus application when creating AppCDS");
-            }
-            throw new PreventFurtherStepsException();
+            InitializationTaskRecorder.preventFurtherRecorderSteps(5,
+                    "Unable to properly shutdown Quarkus application when creating AppCDS",
+                    PreventFurtherStepsException::new);
         }
     }
 }
